refactor menu dropdown component
