+++ src/components/pages/order/index.ts
@@ -3,3 +3,4 @@ export * from './calculate-total'
 export * from './shop-owner'
 export * from './transfer-info'
 export * from './shared-link'
+export * from './menu-dropdown'
